+++ src/main/java/org/gerbil/jsubtitle/ass/StyleSection.java
@@ -6,7 +6,7 @@ public interface StyleSection extends AssFileSection {
 
     String getFontName();
 
-    int getFontSize();
+    double getFontSize();
 
     String getPrimaryColor();
